requirements.txt
h5py==2.7.0
Keras==2.0.2
numpy==1.12.1
Pillow==4.0.0
scikit-learn==0.18.1
scipy==0.19.0
tensorflow==1.0.1
#Theano==0.9.0
